pyproject.toml
[tool.poetry]
name = "azure-ai-102"
version = "0.1.0"
description = "Lab files for Azure AI-102 AI Engineer"
authors = ["Malte Hedderich <github@hedderich.pro>"]
readme = "README.md"
packages = [{include = "azure_ai_102"}]

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.3.2"
requests = "^2.31.0"
matplotlib = "^3.7.2"
Pillow = "^10.0.0"
numpy = "^1.25.1"
azure-ai-textanalytics = "5.3.0"
azure-identity = "1.5.0"
azure-keyvault-secrets = "4.2.0"
azure-cognitiveservices-speech = "1.28.0"
azure-ai-language-conversations = "^1.1.0"
python-dateutil = "^2.8.2"
botbuilder-core = "^4.14.4"
asyncio = "^3.4.3"
aiohttp = "^3.8.5"
cookiecutter = "1.7.0"
azure-cognitiveservices-vision-customvision = "3.1.0"
azure-cognitiveservices-vision-computervision = "0.7.0"
azure-cognitiveservices-vision-face = "0.4.1"
azure-ai-formrecognizer = "3.0.0"
tabulate = "^0.9.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.5"
python-dotenv = "^1.0.0"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
